First TtoDouble commit.
